Create README.md
Hello world, 😊  I work in industrial automation.  We are on the front-end of some wonderful things in the Industrial Internet of Things.  I have a passion for developing streamlined solutions to harvest data, leverage data lakes, databases, and data warehouses and then apply machine learning and artificial intelligence to use the data as a resource.  Using data as a resource means it has value. The value of data is when it is applied back into the ISA 95 stack to intelligently affect ERP, MES, SCADA and other systems.  Many will claim they can do this IIoT, and they can, but the question is will it scale without exponential scaling of price?  Is it open architecture?  Is it lightweight?  Does it conform to current industry best practices when it comes to security?  In order to be successful with IIoT across the whole enterprise, the answer to these questions must be yes to all of them.  I’m on a journey to prove just this!